+++ egui/src/containers/mod.rs
@@ -6,12 +6,20 @@ pub(crate) mod area;
 pub(crate) mod collapsing_header;
 mod combo_box;
 pub(crate) mod frame;
+pub(crate) mod panel;
 pub(crate) mod popup;
 pub(crate) mod resize;
 pub(crate) mod scroll_area;
 pub(crate) mod window;
 
 pub use {
-    area::Area, collapsing_header::*, combo_box::*, frame::Frame, popup::*, resize::Resize,
-    scroll_area::ScrollArea, window::Window,
+    area::Area,
+    collapsing_header::*,
+    combo_box::*,
+    frame::Frame,
+    panel::{SidePanel, TopPanel},
+    popup::*,
+    resize::Resize,
+    scroll_area::ScrollArea,
+    window::Window,
 };