+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 * Refactored the interface for `egui::app::App`
 * Demo App: Add slider to scale all of Egui
 * Windows are now constrained to the screen
-* Panels: you can now add side panels using `Context::panel_left` and `Context::panel_top`.
+* Panels: you can now create panels using `SidePanel` and `TopPanel`.
 * Fix a bug where some regions would slowly grow for non-integral scales (`pixels_per_point`).
 
 ## 0.2.0 - 2020-10-10